Silly note draft of Report
Put in the sections and subsections of my report, along with initial
notes to re-write and build on. Placeholders for images and citations.
Already at around 4-5 pages just with ramblings. Not a bad start though.
Next step is to write the Into and the Basic Features section in detail
and with images.