Removed the camera min/max distance limitations
Since the camera can freely move and change its focus point, it was
useless to have these limitations upon the camera. It would just lead to
people moving freely, zooming in, moving back and then wondering why
they can't zoom anymore. Best to just allow them the freedom, even if it
means to can go inside of the Sun or outside of the Star map.